Initial Function Creation: I created the CRUD (Create, Read, Update, Delete) operations for the User model. This was achieved through the creation of the functions createUser, getAllUsers, updateUser, and deleteUser. This fundamental step has enabled efficient interaction with the User data in our MongoDB database.
Schema Refactoring and Function Update: In response to changes in the project requirements, I refined the User schema and updated the createUser and updateUser functions to reflect these changes. Key modifications in the schema include the removal of profile picture, social media, and subscription fields. I've also added a role field to differentiate between user types (such as admin, seller, and customer), and a purchases field to keep track of the user's total expenditure and purchase count.

The changes to the createUser and updateUser functions were implemented to handle these new fields appropriately. As part of this, I've ensured that when a new user is created, initial values are set for the totalSpent and count fields within purchases.

These updates enhance the functionality of the User model, making it more adaptable for our e-commerce context. I've ensured that the createUser and updateUser functions are now fully capable of managing the new fields, thereby maintaining accurate data management for our users.